Adapt upstream UI to ModernUI 3.12 layout and color interfaces

+++ common/src/main/java/indi/mopelotus/musichud/client/ui/hud/HudEditOverlayView.java
@@ -291,7 +291,7 @@ public class HudEditOverlayView extends View {
     }
 
     private void drawHandleMarker(Canvas canvas, Handle handle, float cx, float cy) {
-        long color = Theme.HUD_HANDLE_STATES.getColorForState(handleState(handle),
+        int color = Theme.HUD_HANDLE_STATES.getColorForState(handleState(handle),
                 Theme.HUD_HANDLE_STATES.getDefaultColor());
         handlePaint.setColor(color);
         switch (handle) {
